feat: Team CRD (#2691)

+++ cmd/server.go
@@ -144,6 +144,15 @@ func launchKopper(ctx context.Context) {
 		shutdown.ShutdownAndExit(1, fmt.Sprintf("Unable to create controller for Scope: %v", err))
 	}
 
+	if _, err := kopper.SetupReconciler(ctx, mgr,
+		db.PersistTeamFromCRD,
+		db.DeleteTeam,
+		db.DeleteStaleTeam,
+		"team.mission-control.flanksource.com",
+	); err != nil {
+		shutdown.ShutdownAndExit(1, fmt.Sprintf("Unable to create controller for Team: %v", err))
+	}
+
 	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
 		shutdown.ShutdownAndExit(1, fmt.Sprintf("error running controller manager: %v", err))
 	}
@@ -301,6 +310,9 @@ func tableUpdatesHandler(ctx context.Context) {
 
 			switch tgOperation {
 			case TGOPDelete:
+				if rbac.Enforcer() == nil {
+					break
+				}
 				auth.InvalidateRLSCacheForUser(personID)
 				if err := rbac.DeleteRoleForUser(personID, teamID); err != nil {
 					ctx.Errorf("failed to delete team(%s)->user(%s) rbac policy: %v", teamID, personID, err)
@@ -309,6 +321,9 @@ func tableUpdatesHandler(ctx context.Context) {
 				}
 
 			case TGOPInsert, TGOPUpdate:
+				if rbac.Enforcer() == nil {
+					break
+				}
 				auth.InvalidateRLSCacheForUser(personID)
 				if err := rbac.AddRoleForUser(personID, teamID); err != nil {
 					ctx.Errorf("failed to add team(%s)->user(%s) rbac policy: %v", teamID, personID, err)
